+++ testdata/embedded_flatten_simple/input.go
@@ -1,0 +1,12 @@
+package embeddedflattensimple
+
+type Base struct {
+	ID string `json:"id"`
+	N  int    `json:"n"`
+}
+
+//gents:export
+type Foo struct {
+	Base
+	Name string `json:"name"`
+}